fastapi_next/identity/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Meysam Feghhi <me.feghhi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
google-auth = "^2.35.0"
requests = "^2.32.3"
pyjwt = "^2.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.29"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
dev.cmd = "python -B app/main.py"
start.cmd = "python app/main.py"
console.cmd = "python -B app/console.py"
db-create.script = "app.database:create_database"
db-drop.script = "app.database:drop_database"
db-migrate.cmd = "alembic upgrade head"
mypy.cmd = "mypy ."
test.cmd = "pytest"

[tool.mypy]
strict = true